server/requirements.txt
flask
markdown
requests
pyyaml